@@ -1,43 +1,49 @@
 {
+  "name": "@pmndrs/docs",
+  "version": "0.1.0",
+  "private": true,
+  "type": "module",
+  "bin": {
+    "build": "bin/build.mjs"
+  },
   "devDependencies": {
-    "@types/node": "^18.6.4",
-    "@types/react": "^18.3.0",
+    "@types/minimist": "^1.2.5",
+    "@types/node": "^18.19.44",
+    "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
-    "autoprefixer": "^10.4.8",
-    "eslint": "^8.21.0",
+    "autoprefixer": "^10.4.20",
+    "eslint": "^8.57.0",
     "eslint-config-next": "^14.2.5",
-    "postcss": "^8.4.16",
-    "prettier": "^2.7.1",
-    "sharp": "^0.30.7",
-    "tailwindcss": "^3.1.8",
+    "minimist": "^1.2.8",
+    "postcss": "^8.4.41",
+    "prettier": "^2.8.8",
+    "semantic-release": "^24.0.0",
+    "tailwindcss": "^3.4.9",
     "typescript": "^5.5.4"
   },
   "dependencies": {
-    "@headlessui/react": "^1.6.6",
-    "@tailwindcss/aspect-ratio": "^0.4.0",
-    "@tailwindcss/typography": "^0.5.4",
-    "clsx": "^1.2.1",
-    "gray-matter": "^4.0.2",
-    "isomorphic-git": "^1.19.1",
-    "match-sorter": "^6.3.1",
-    "memfs": "^3.4.7",
+    "@tailwindcss/aspect-ratio": "^0.4.2",
+    "@tailwindcss/typography": "^0.5.14",
+    "clsx": "^2.1.1",
+    "gray-matter": "^4.0.3",
+    "match-sorter": "^6.3.4",
     "next": "^14.2.5",
-    "next-mdx-remote": "^4.1.0",
+    "next-mdx-remote": "^4.4.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-icons": "^4.12.0",
-    "rehype-prism-plus": "^1.5.0",
+    "rehype-prism-plus": "^1.6.3",
     "remark-gfm": "^3.0.1"
   },
   "scripts": {
-    "dev": "next",
+    "dev": "NODE_OPTIONS='--inspect' next",
     "start": "next start",
     "build": "next build",
     "format": "prettier -w src/",
     "lint": "eslint src/**/*.{ts,tsx} && prettier . --check"
   },
   "engines": {
-    "node": ">=18"
+    "node": ">=20"
   },
-  "packageManager": "yarn@1.22.19"
+  "packageManager": "npm@10.8.2"
 }
