@@ -1,23 +1,23 @@
 import * as React from 'react'
-import { useRouter } from 'next/router'
+import { useRouter } from 'next/navigation'
 import { matchSorter } from 'match-sorter'
 
-import { useCSB } from 'hooks/useCSB'
-import { useDocs } from 'hooks/useDocs'
+import { useDocs } from '@/app/[...slug]/DocsContext'
 
 import SearchModal from './SearchModal'
 import type { SearchResult } from './SearchItem'
-import { escape } from 'utils/text'
+import { escape } from '@/utils/text'
 
 export interface SearchModalContainerProps {
   onClose: React.MouseEventHandler<HTMLButtonElement>
 }
 
 export const SearchModalContainer = ({ onClose }: SearchModalContainerProps) => {
-  const router = useRouter()
-  const boxes = useCSB()
-  const docs = useDocs()
-  const [lib] = router.query.slug as string[]
+  // const router = useRouter()
+  // const boxes = useCSB()
+  const { docs } = useDocs()
+  console.log('docs', docs)
+  // const [lib] = router.query.slug as string[]
   const [query, setQuery] = React.useState('')
   const deferredQuery = React.useDeferredValue(query)
   const [results, setResults] = React.useState<SearchResult[]>([])
@@ -25,6 +25,7 @@ export const SearchModalContainer = ({ onClose }: SearchModalContainerProps) => 
   React.useEffect(() => {
     React.startTransition(() => {
       if (!deferredQuery) return setResults([])
+      // console.log('deferredQuery', deferredQuery)
 
       // Get length of matched text in result
       const relevanceOf = (result: SearchResult) =>
@@ -32,20 +33,23 @@ export const SearchModalContainer = ({ onClose }: SearchModalContainerProps) => 
         result.title.length
 
       // Search
-      const entries = (docs.flatMap(({ tableOfContents }) => tableOfContents) as SearchResult[])
-        .filter((entry) => entry.description.length > 0)
-        .concat(
-          Object.entries(boxes).flatMap(([id, data]) => ({
-            ...data,
-            label: 'codesandbox.io',
-            description: data.description ?? '',
-            content: data.content ?? '',
-            url: `https://codesandbox.io/s/${id}`,
-            image: data.screenshot_url,
-          }))
-        )
+      let candidateResults = docs.flatMap(
+        ({ tableOfContents }) => tableOfContents
+      ) satisfies SearchResult[]
+      // console.log('candidateResults', candidateResults)
+      candidateResults = candidateResults.filter((entry) => entry.description.length > 0)
+      // .concat(
+      //   Object.entries(boxes).flatMap(([id, data]) => ({
+      //     ...data,
+      //     label: 'codesandbox.io',
+      //     description: data.description ?? '',
+      //     content: data.content ?? '',
+      //     url: `https://codesandbox.io/s/${id}`,
+      //     image: data?.screenshot_url,
+      //   }))
+      // )
 
-      const results = matchSorter(entries, deferredQuery, {
+      const results = matchSorter(candidateResults, deferredQuery, {
         keys: ['title', 'description', 'content'],
         threshold: matchSorter.rankings.CONTAINS,
       })
@@ -56,7 +60,7 @@ export const SearchModalContainer = ({ onClose }: SearchModalContainerProps) => 
 
       setResults(results)
     })
-  }, [boxes, docs, lib, deferredQuery])
+  }, [docs, deferredQuery])
 
   return (
     <SearchModal
